Most of Sending TEbucks

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -3,31 +3,31 @@ package com.techelevator.tenmo.model;
 import java.math.BigDecimal;
 
 public class Account {
-    private int account_id;
-    private int user_id;
+    private long account_id;
+    private long user_id;
     private BigDecimal balance;
 
     public Account() {}
 
-    public Account(int account_id, int user_id, BigDecimal balance) {
+    public Account(long account_id, long user_id, BigDecimal balance) {
         this.account_id = account_id;
         this.user_id = user_id;
         this.balance = balance;
     }
 
-    public int getAccount_id() {
+    public long getAccount_id() {
         return account_id;
     }
 
-    public void setAccount_id(int account_id) {
+    public void setAccount_id(long account_id) {
         this.account_id = account_id;
     }
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
